ATTENTION: Field option "hide" renamed to "hidden". It now as other propose.
"hidden" Instructs the datagrid to pass the field to the deploy class. Is the deploy class the responsible to hide those fields. A new method has been added to help developers handle this. Pass in your deploy class in __contrcut() method $this->_setRemoveHiddenFields(true); and the datagrid will handle both options as "remove".

"remove" the formerly hide option. The field value can still be used for format, decorator, etc, but won't be passed tot he deploy class.